Make result order deterministic (#126)
This makes the order of results in a `Result.*Pool` deterministic so
that the order of the result slice corresponds with the order of tasks
submitted. As an example of why this would be useful, it makes it easy
to rewrite `iter.Map` in terms of `ResultPool`. Additionally, it's a
generally nice and intuitive property to be able to match the index of
the result slice with the index of the input slice.