Reusable pools (#129)
This updates the pool types that collect results and errors to reset on
`Wait` so they are reusable once waited on. Previously, if a pool was
reused, the returned values of `Wait()` would contain the aggregated set
of all previous uses. This wasn't explicitly a guarantee of the library
before, but it does make it operate more like `sync.WaitGroup` and it's
easy to do, so I think it's a positive change.